Attempt to get supplied sugarcane mask code working (Mega_Sugar_Masks.py)
All the supplied code was rolled into a single program, because I was
fed up trying to make them see each other.

These masks are built differently at the moment -- anything that is "IN"
the area is coloured black 0/0/0, the rest is taken from an image of the
fields, so you can see the rest.  This is different to how the phase 1
masks worked.  We either have to adapt to the new mask format, or tweak
this program to produce the old program.  I'm leaning towards adapting
to the new format, because it's useful to be able to see that the
non-sugar areas really aren't fields.